+++ packages/toolpad-app/src/createRuntimeState.tsx
@@ -1,36 +1,6 @@
-import { codeFrameColumns } from '@babel/code-frame';
-import { transform, TransformResult } from 'sucrase';
 import * as appDom from './appDom';
+import compileModule from './compileModule';
 import { CompiledModule, RuntimeState } from './types';
-import { errorFrom } from './utils/errors';
-import { findImports, isAbsoluteUrl } from './utils/strings';
-
-export function compileModule(src: string, filename: string): CompiledModule {
-  const urlImports = findImports(src).filter((spec) => isAbsoluteUrl(spec));
-
-  let compiled: TransformResult;
-
-  try {
-    compiled = transform(src, {
-      transforms: ['jsx', 'typescript', 'imports'],
-      filePath: filename,
-      jsxRuntime: 'classic',
-    });
-  } catch (rawError) {
-    const error = errorFrom(rawError);
-    if ((error as any).loc) {
-      error.message = [error.message, codeFrameColumns(src, { start: (error as any).loc })].join(
-        '\n\n',
-      );
-    }
-    return { error };
-  }
-
-  return {
-    ...compiled,
-    urlImports,
-  };
-}
 
 function compileModules(dom: appDom.AppDom): Record<string, CompiledModule> {
   const result: Record<string, CompiledModule> = {};
